Combat Tick Counter refactor
